ARC: mem: introduce memory read function
With this commit we introduce exported arc_mem_read function
and arc_mem_is_slow_memory,arc_mem_read_block supporting
functions.

The code without significant changes was taken from:
https://github.com/foss-for-synopsys-dwc-arc-processors/openocd/blob/arc-2018.09-release/src/target/arc_mem.*

The dcache_flush call is commented because this function is not introduced yet.

We still keep the code for big-endiannes setups for futher to be introduced BE setups.

Signed-off-by: Evgeniy Didin <didin@synopsys.com>